+++ Changes.md
@@ -1,3 +1,4 @@
+* #716 - Add command line option to start webserver immediately
 * #720 - Add performance score report for finalist judging
 * #717 - Confirm closing the launcher when the web server is running
 * #730 - Add table for each goal group within a subjective category to the subjective score report